requirements.txt
pandas
numpy
scikit-learn
torch
transformers
spacy
lime
anchor-exp
gensim
textblob
tabulate
matplotlib
datasets
shap
ferret
